Add all parameters with showing results

+++ Dietician/Dietician/Models/ParameterResults.cs
@@ -4,9 +4,9 @@
     {
         public double Bmi { get; set; }
         public string BmiLabel { get; set; }
-        public int Cmp { get; set; }
-        public int Protein { get; set; } 
-        public int Carbohydrate { get; set; }
+        public double Cmp { get; set; }
+        public double Protein { get; set; } 
+        public double Carbohydrate { get; set; }
         public int Fat { get; set; }
         public double Whr { get; set; }
         public string WhrLabel { get; set; }
